refactor: relocate tokens to src/shared/tokens
