+++ src/shared/tokens/data/color.json
